Remove Flask based frontend
