+++ packit_dashboard/utils.py
@@ -1,4 +1,17 @@
-from packit_dashboard.packagewide_utils import return_json
+import requests
+import json
+
+
+# Common utility functions used in multiple files in the packit_dashboard package
+# Returns python parsable json object from URL
+def return_json(url, method="GET", **kwargs):
+    response = requests.request(method=method, url=url, **kwargs)
+    output = None
+    try:
+        output = json.loads(response.content)
+    except Exception:
+        pass
+    return output
 
 
 # Takes a base url, and gets json files from multiple pages from that url
