--- templates/builds.html
@@ -1,99 +1,0 @@
-<h2>Recent Builds</h2>
-<table
-  class="pf-c-table pf-m-grid-lg"
-  role="grid"
-  aria-label="Builds Table"
-  id="builds-table"
->
-  <thead>
-    <tr role="row">
-      <th class="pf-c-table__sort " role="columnheader" scope="col">
-        <button class="pf-c-button pf-m-plain" type="button">
-          Project
-          <span class="pf-c-table__sort-indicator">
-            <i class="fas fa-arrows-alt-v"></i>
-          </span>
-        </button>
-      </th>
-      <th
-        class="pf-c-table__sort "
-        role="columnheader"
-        aria-sort="none"
-        scope="col"
-      >
-        <button class="pf-c-button pf-m-plain" type="button">
-          Owner
-          <span class="pf-c-table__sort-indicator">
-            <i class="fas fa-arrows-alt-v"></i>
-          </span>
-        </button>
-      </th>
-      <th
-        class="pf-c-table__sort "
-        role="columnheader"
-        aria-sort="none"
-        scope="col"
-      >
-        <button class="pf-c-button pf-m-plain" type="button">
-          Status
-          <span class="pf-c-table__sort-indicator">
-            <i class="fas fa-arrows-alt-v"></i>
-          </span>
-        </button>
-      </th>
-      <th
-        class="pf-c-table__sort sort-disabled"
-        role="columnheader"
-        aria-sort="none"
-        scope="col"
-      >
-        <button class="pf-c-button pf-m-plain" type="button">Chroots</button>
-      </th>
-      <th
-        class="pf-c-table__sort "
-        role="columnheader"
-        aria-sort="none"
-        scope="col"
-      >
-        <button class="pf-c-button pf-m-plain" type="button">
-          Build ID
-          <span class="pf-c-table__sort-indicator">
-            <i class="fas fa-arrows-alt-v"></i>
-          </span>
-        </button>
-      </th>
-    </tr>
-  </thead>
-  <tbody role="rowgroup">
-    {% for build in builds_list %}
-    <tr role="row">
-      <td role="cell" data-label="Project">{{ build["project"] }}</td>
-      <td role="cell" data-label="Owner">{{ build["owner"] }}</td>
-      <td role="cell" data-label="Status">{{ build["status"] }}</td>
-      <td role="cell" data-label="Chroots">
-        {% for os in build["chroots"] %}
-        <li>{{ os }}</li>
-        {% endfor %}
-      </td>
-      <td role="cell" data-label="Build ID">{{ build["build_id"] }}</td>
-    </tr>
-    {% endfor %}
-  </tbody>
-</table>
-
-<script src="{{ url_for('node_modules', filename='datatables.net/js/jquery.dataTables.min.js') }}"></script>
-<script type="text/javascript">
-  $(document).ready(function() {
-    $("#builds-table").DataTable({
-      paging: false,
-      searching: false,
-      info: false
-    });
-  });
-</script>
-
-<style>
-  .sort-disabled {
-    pointer-events: none;
-  }
-</style>
