--- templates/information.html
@@ -1,8 +1,0 @@
-<h2>About Packit Service</h2>
-
-<strong>Packit-as-a-Service</strong> is a GitHub App that helps you continuously
-ensure that your project work in Fedora OS.
-
-<h2>Documentation</h2>
-Documentation is available
-<a href="https://packit.dev/packit-as-a-service/">online</a>
