--- templates/main_frame.html
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-
-    <link
-      rel="stylesheet"
-      type="text/css"
-      href="{{ url_for('node_modules', filename='@patternfly/patternfly/patternfly.css') }}"
-    />
-
-    <!-- Scripts at the top so that js based components do not error out -->
-    <script src="{{ url_for('node_modules', filename='jquery/dist/jquery.min.js') }}"></script>
-    <script src="{{ url_for('node_modules', filename='patternfly/dist/js/patternfly.js') }}"></script>
-  </head>
-
-  <body>
-    <div class="pf-c-page" id="page-layout-horizontal-nav">
-      <header role="banner" class="pf-c-page__header">
-        <div class="pf-c-page__header-brand">
-          <a
-            class="pf-c-page__header-brand-link"
-            href="https://github.com/packit-service/packit-service"
-          >
-            <img
-              class="pf-c-brand"
-              src="https://avatars2.githubusercontent.com/u/46870917?s=400&u=d55055f0ee529d7b45dbf67f464735527becca77&v=4"
-              alt="Packit Service"
-            />
-          </a>
-        </div>
-        <div class="pf-c-page__header-nav">
-          <nav
-            class="pf-c-nav pf-m-end"
-            id="page-layout-horizontal-nav-horizontal-nav"
-            aria-label="Global"
-          >
-            <ul class="pf-c-nav__horizontal-list">
-              <li class="pf-c-nav__item">
-                <a href="{{ url_for('about.status') }}" class="pf-c-nav__link">
-                  Service Status
-                </a>
-              </li>
-              <li class="pf-c-nav__item">
-                <a
-                  href="{{ url_for('projects.projects_page') }}"
-                  class="pf-c-nav__link"
-                >
-                  Projects
-                </a>
-              </li>
-              <li class="pf-c-nav__item">
-                <a
-                  href="{{ url_for('builds.builds_page') }}"
-                  class="pf-c-nav__link"
-                >
-                  Recent Builds
-                </a>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </header>
-
-      <main
-        role="main"
-        class="pf-c-page__main"
-        tabindex="-1"
-        id="main-content-page-layout-horizontal-nav"
-      >
-        <section class="pf-c-page__main-section pf-m-light">
-          <div class="pf-c-content">
-            <h1>{{header}}</h1>
-          </div>
-        </section>
-        <section class="pf-c-page__main-section pf-m-light">
-          <div class="pf-c-content">
-            {{ content|safe }}
-          </div>
-        </section>
-      </main>
-    </div>
-  </body>
-</html>
