--- templates/project.html
@@ -1,12 +1,0 @@
-<div class="pf-c-card">
-  <div class="pf-c-card__header pf-c-title pf-m-md">
-    <i class="fas {{ icon }}" aria-hidden="true"></i>
-    {{ header}}
-  </div>
-  <div class="pf-c-card__body">
-    {{ text|safe }}
-  </div>
-  <div class="pf-c-card__footer">
-    <a href="{{ copr_link }}">Copr project page</a>
-  </div>
-</div>
