--- templates/projects.html
@@ -1,47 +1,0 @@
-<ul
-  class="pf-c-data-list"
-  role="list"
-  aria-label="Basic data list example"
-  id="data-list-basic"
->
-  <li class="pf-c-data-list__item" aria-labelledby="data-list-basic-item1">
-    <div class="pf-c-data-list__item-row">
-      <div class="pf-c-data-list__item-content">
-        <div class="pf-c-data-list__cell">
-          <span id="data-list-basic-item1">
-            <h2>Total projects: {{ counter }}</h2>
-            (Projects which ever used (i.e. were building their PR code in Copr
-            via) Packit service)
-            <div>
-              <ul>
-                {% for project in all_projects %}
-                <li>{{ project }}</li>
-                {% endfor %}
-              </ul>
-            </div>
-          </span>
-        </div>
-      </div>
-    </div>
-  </li>
-  <li class="pf-c-data-list__item" aria-labelledby="data-list-basic-item2">
-    <div class="pf-c-data-list__item-row">
-      <div class="pf-c-data-list__item-content">
-        <div class="pf-c-data-list__cell">
-          <span id="data-list-basic-item2">
-            <h2>Successful projects: {{ counter_s }}</h2>
-            (Projects which ever successfully used (i.e. successfully built
-            their PR code in Copr in all targets via) Packit service)
-            <div>
-              <ul>
-                {% for project in successful_projects %}
-                <li>{{ project }}</li>
-                {% endfor %}
-              </ul>
-            </div>
-          </span>
-        </div>
-      </div>
-    </div>
-  </li>
-</ul>
