--- templates/status.html
@@ -1,7 +1,0 @@
-<div class="pf-c-card">
-  <div class="pf-c-card__header pf-c-title pf-m-md">
-    <i class="fas {{ icon }}" aria-hidden="true"></i>
-    {{ header}}
-  </div>
-  <div class="pf-c-card__body">{{ text|safe }}</div>
-</div>
